+++ lib/src/core/app/di.dart
@@ -12,7 +12,6 @@ import '../exceptions/_exceptions.dart';
 
 final getIt = GetIt.instance;
 
-// TODO watchout disposing them.
 /// Registering dependency injections for core modules.
 Future<void> initCoreAppModule() async {
   // Randomization