+++ lib/src/util/context_shortcut.dart
@@ -2,7 +2,7 @@
 import 'package:flutter/material.dart';
 
 import '../core/_core.dart';
-import 'navigation.dart';
+import '../core/routing/navigation.dart';
 
 // ! [WARNING]
 // ! Everything here must only be used one widget below Material widget.